parsers: moved create_style_element out of convert_payload_to_html
